requirements.txt
sentence-transformers
tqdm
faiss-gpu
tabulate
sentencepiece
sacrebleu
tensorboard
kenlm
matplotlib